// packages/components/src/optional/CustomComponents/types.ts
export interface CustomComponentsTranslations {
  seeing: string
  hearing: string
  mobility: string
  self_care: string
  communication: string
  no_difficulty: string
  your_gender: string
  your_gender_identity: string
  disability_question: string
  undisclosed_religion: string
  religion_question: string
  encyclopedia_version_question: string
  religion_perspective_heading: string
  religion_perspective_content: string
  islamic_perspective_heading: string
  islamic_perspective_content: string
}